chore: Remove no-check flag from compile task
This flag was initially included because Deno was trying to check the types of the .ts files in the template directory, which is included in the compilation so that it can be copied but isn't intended to be executable. The compilation would fail if the Purescript module hadn't been built.

The flag is no longer necessary because the .ts files have had a .template extension added, which prevents Deno from checking the modules.